+++ src/main/java/atom/marvelores/MarvelOres.java
@@ -1,8 +1,11 @@
 package atom.marvelores;
 
 import atom.marvelores.block.ModBlocks;
+import atom.marvelores.block.entity.ModBlockEntities;
 import atom.marvelores.creativemodetab.ModCreativeModeTabs;
+import atom.marvelores.custom.ModMenuTypes;
 import atom.marvelores.item.ModItems;
+import atom.marvelores.recipe.ModRecipes;
 import atom.marvelores.worldgen.gen.ModWorldGeneration;
 import net.fabricmc.api.ModInitializer;
 
@@ -21,5 +24,10 @@ public class MarvelOres implements ModInitializer {
 		ModBlocks.registerModBlocks();
 
 		ModWorldGeneration.generateModWorldGen();
+
+		ModBlockEntities.registerBlockEntities();
+
+		ModMenuTypes.registerModMenuTypes();
+		ModRecipes.registerModRecipes();
 	}
 }