+++ src/main/java/atom/marvelores/block/custom/CrusherBlock.java
@@ -1,0 +1,94 @@
+package atom.marvelores.block.custom;
+
+
+import atom.marvelores.block.entity.ModBlockEntities;
+import atom.marvelores.block.entity.custom.CrusherBlockEntity;
+import com.mojang.serialization.MapCodec;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.core.particles.ItemParticleOption;
+import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.BaseEntityBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.phys.BlockHitResult;
+import org.jetbrains.annotations.Nullable;
+
+public class CrusherBlock extends BaseEntityBlock {
+    public static final EnumProperty<Direction> FACING = BlockStateProperties.HORIZONTAL_FACING;
+    public static final BooleanProperty LIT = BlockStateProperties.LIT;
+    public static final MapCodec<CrusherBlock> CODEC = simpleCodec(CrusherBlock::new);
+
+
+    public CrusherBlock(Properties properties) {
+        super(properties);
+    }
+
+    /* FACING */
+    @Override
+    public @Nullable BlockState getStateForPlacement(BlockPlaceContext context) {
+        return this.defaultBlockState().setValue(FACING, context.getHorizontalDirection().getOpposite()).setValue(LIT, false);
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        builder.add(FACING, LIT);
+    }
+
+    /* BLOCK ENTITY */
+    @Override
+    protected MapCodec<? extends BaseEntityBlock> codec() {
+        return CODEC;
+    }
+
+    @Override
+    public @Nullable BlockEntity newBlockEntity(BlockPos worldPosition, BlockState blockState) {
+        return new CrusherBlockEntity(worldPosition, blockState);
+    }
+
+    @Override
+    public void playerDestroy(Level level, Player player, BlockPos pos, BlockState state,
+                              @Nullable BlockEntity blockEntity, ItemStack destroyedWith) {
+        if(level.getBlockEntity(pos) instanceof CrusherBlockEntity CrusherBlockEntity) {
+            CrusherBlockEntity.drops();
+        }
+        super.playerDestroy(level, player, pos, state, blockEntity, destroyedWith);
+    }
+
+    @Override
+    protected InteractionResult useWithoutItem(BlockState state, Level level, BlockPos pos,
+                                               Player player, BlockHitResult hitResult) {
+        if(!level.isClientSide()) {
+            if(level.getBlockEntity(pos) instanceof CrusherBlockEntity CrusherBlockEntity) {
+                player.openMenu(CrusherBlockEntity);
+            }
+        }
+        return InteractionResult.SUCCESS;
+    }
+
+    @Override
+    public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState blockState,
+                                                                            BlockEntityType<T> type) {
+        if(level.isClientSide()) {
+            return null;
+        }
+
+        return createTickerHelper(type, ModBlockEntities.CRUSHER_BE,
+                (level1, pos, state, entity) -> entity.tick(level1, pos, state));
+    }
+}