+++ src/main/java/atom/marvelores/datagen/ModBlockTagProvider.java
@@ -19,13 +19,15 @@ public class ModBlockTagProvider extends FabricTagsProvider.BlockTagsProvider {
                 .add(ModBlocks.VIBRANIUM_BLOCK)
                 .add(ModBlocks.RAW_VIBRANIUM_BLOCK)
                 .add(ModBlocks.VIBRANIUM_ORE)
-                .add(ModBlocks.VIBRANIUM_DEEPSLATE_ORE);
+                .add(ModBlocks.VIBRANIUM_DEEPSLATE_ORE)
+                .add(ModBlocks.CRUSHER);
 
 
         valueLookupBuilder(BlockTags.NEEDS_DIAMOND_TOOL)
                 .add(ModBlocks.VIBRANIUM_BLOCK)
                 .add(ModBlocks.RAW_VIBRANIUM_BLOCK)
                 .add(ModBlocks.VIBRANIUM_ORE)
-                .add(ModBlocks.VIBRANIUM_DEEPSLATE_ORE);
+                .add(ModBlocks.VIBRANIUM_DEEPSLATE_ORE)
+                .add(ModBlocks.CRUSHER);
     }
 }