+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## VIBRANIUM UPDATE !!
 
-+ Full armor effects !
-
++ Add Vibranium Stick
++ Add Vibranium Core
++ Add Vibranium Powder
++ Add Crusher
++ Change some craft
++ 
 _Add a lot more coming soon !_